Break seed file apart into multiple files
