+++ Gemfile
@@ -58,8 +58,9 @@ gem 'sidekiq'
 gem 'dalli'
 
 group :development do
-  gem 'perf_check'
-  gem 'annotate'
+  gem 'perf_check', require: false
+  gem 'annotate', require: false
+  gem 'faker', require: false
   # Available "channels" of rubocop for code climate:
   # https://github.com/codeclimate/codeclimate-rubocop/branches/all?utf8=✓&query=channel%2Frubocop
   gem 'rubocop', '1.4.1', require: false # synced to .codeclimate.yml
